backup/v1/requirements.txt
starlette
fastapi 
uvicorn
requests
google-generativeai 
python-dotenv
os